Moved all wrap related things to their own submodule.
